Refactor environment variables, add .gitignore, and enhance controller functionality
- Updated .env file to include API_PORT.
- Added .gitignore to exclude .env files.
- Changed DB_URL format in Makefile.
- Renamed obraControle to obraController for consistency.
- Added GetObraById and GetDiariosByObraId methods in respective controllers.
- Implemented error handling for ID parameters in controller methods.
- Updated database connection settings in docker-compose.yml to use environment variables.
- Enhanced Diario and Usuario models to include updated_at field.
- Modified SQL migrations to add updated_at fields in obra and diario_obra tables.

+++ internal/services/usuario.go
@@ -16,7 +16,7 @@ func NewUsuarioService(connection *sql.DB) UsuarioServices {
 	}
 }
 
-func (pr *UsuarioServices) CreateUsuario(user models.Usuario, senhaHash []byte) (int, error) {
+func (pr *UsuarioServices) CreateUsuario(user models.Usuario, senhaHash []byte) (int64, error) {
 	var id int
 
 	query := `
@@ -52,7 +52,7 @@ func (pr *UsuarioServices) CreateUsuario(user models.Usuario, senhaHash []byte) 
 		return 0, err
 	}
 
-	return id, nil
+	return int64(id), nil
 }
 
 func (pr *UsuarioServices) GetUsuarios() ([]models.Usuario, error) {
@@ -126,3 +126,9 @@ func (pr UsuarioServices) GetUsuarioById(id int) (models.Usuario, error) {
 	return usuario, nil
 
 }
+
+// func (pr UsuarioServices) PatchUsuarioById(id int, usuario models.Usuario) (models.Usuario, error) {
+// 	var id int
+
+// 	query  := `update usuario`
+// }
